% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BER.R
\name{BER}
\alias{BER}
\title{Batch Effect Removal}
\usage{
BER(X1, X2)
}
\arguments{
\item{X1}{matrix of gene expression from batch 1 (excluding References). Rows are samples cols are genes.}

\item{X2}{matrix of gene expression from batch 2 (excluding References). Rows are samples cols are genes.}
}
\value{
matrix of log normalized data in the same format but without reference genes.
}
\description{
This function adjusts the gene expression level between to batches using ber with bagging (1000 sims).
}
\note{
the column names of X1 and X2 have to match.
}

