% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCplot.R
\name{CCplot}
\alias{CCplot}
\title{Concordance Correlation Plot}
\usage{
CCplot(method1, method2, Ptype = "None", metrics = FALSE, xlabel = "",
  ylabel = "", title = "", subtitle = "", xrange = NULL,
  yrange = NULL, MArange = c(-3.5, 5.5))
}
\arguments{
\item{method1}{measurements obtained in batch 1 or using method 1}

\item{method2}{measurements obtained in batch 2 or using method 2}

\item{Ptype}{type of plot to be outputted c("scatter", "MAplot")}

\item{metrics}{if return metrics is set to \code{TRUE} (default is \code{FALSE}), returns Rc, Ca and R2}

\item{xlabel}{label to be used for x axis}

\item{ylabel}{label to be used for y axis}

\item{title}{title for the main plot}

\item{subtitle}{subtitle if requested}

\item{xrange}{range of x axis}

\item{yrange}{range of y axis}

\item{MArange}{MA range}
}
\description{
Plotting function for Reliability
}
\author{
Aline Talhouk
}

