% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DSC.R
\name{DSC}
\alias{DSC}
\title{Compute the Dispersion Separability Criterion (DSC) which indicates how similar two data sets are by considering the between data scatter matrix and the withing data scatter matrix}
\usage{
DSC(data, Y, permut = 10)
}
\arguments{
\item{data}{is a combined data set}

\item{Y}{a vector indicating the different batches}

\item{permut,}{the number of permutations to compute a pvalue for DSC using a permutation test.}
}
\description{
Compute the Dispersion Separability Criterion (DSC) which indicates how similar two data sets are by considering the between data scatter matrix and the withing data scatter matrix
}
\author{
Aline Talhouk
}

