% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycombat_p.R
\name{mycombat_p}
\alias{mycombat_p}
\title{Custom combat_p}
\usage{
mycombat_p(Y, b, prior.plots = F)
}
\arguments{
\item{Y}{A data matrix}

\item{b}{Vector indicating batches}

\item{prior.plots}{If \code{TRUE}, prior plots to compare kernel density estimate and
parametric estimate are produced.}
}
\description{
Based on combat_p function from package \code{ber}
}
\author{
Marco Giordan, Aline Talhouk
}
