% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratioMethod.R
\name{ratioMethod}
\alias{ratioMethod}
\title{Batch adjustment by taking the ratio of a reference sample}
\usage{
ratioMethod(Y2, R1, R2)
}
\arguments{
\item{Y2}{is the data run in the second batch, samples are rows and genes are columns}

\item{R1}{is the reference data run in the first batch}

\item{R2}{is the reference data run in the second batch}
}
\value{
Y2new is the Y2 data adjusted to batch 1
}
\description{
Batch adjustment by taking the ratio of a reference sample
}
\author{
Aline Talhouk
}

