% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scattermat.R
\name{scattermat}
\alias{scattermat}
\title{Function that computes the scatter matrix}
\usage{
scattermat(data, Y)
}
\arguments{
\item{data}{data matrix}

\item{Y}{is a vector of classes}
}
\value{
A list with the following elements
\item{B}{between class scatter matrix}
\item{W}{within class scatter matrix}
}
\description{
Function that computes the scatter matrix
}
\author{
Aline Talhouk
}

