% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMC.R
\name{BMC}
\alias{BMC}
\title{Batch Mean Centering}
\usage{
BMC(X1, X2)
}
\arguments{
\item{X1}{matrix of gene expression from batch 1 (excluding References). Rows are samples cols are genes.}

\item{X2}{matrix of gene expression from batch 2 (excluding References). Rows are samples cols are genes.}
}
\value{
a combined matrix of log normalized data in the same format but without reference genes.
}
\description{
This functions adjust the gene expression level between to batches by removing the average expression of each gene across cases in each batch, and adding back the average log expression from batch 1
}
\note{
the column names of X1 and X2 have to match.
}

