% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCplot.R
\name{CCplot}
\alias{CCplot}
\title{Plotting function for Reliability}
\usage{
CCplot(method1, method2, Ptype = "none", metrics = FALSE, xlabel = "",
  ylabel = "", ttile = "", subtitle = "")
}
\arguments{
\item{method1}{is the measurements obtained in batch 1 or using method 1}

\item{method2}{is the measurements obtained in batch 2 or using method 2}

\item{Ptype}{is the type of plot to be outputted c("scatter","MAplot")}

\item{xlabel}{is the label to be used for x axis}

\item{ylabel}{is the label to be used for y axis}

\item{subtitle}{is the subtitle if requested}

\item{Metrics}{if return metrics is set to true(default is FALSE) returns Rc, Ca and R2}

\item{ttitle}{the title for the main plot}
}
\description{
Plotting function for Reliability
}

