% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DSC.R
\name{DSC}
\alias{DSC}
\title{Dispersion Separability Criterion (DSC)}
\usage{
DSC(data, Y, permut = 10, digits.d = 3, digits.p = floor(log10(permut)) +
  1)
}
\arguments{
\item{data}{a combined data set}

\item{Y}{vector indicating the two different batches}

\item{permut}{number of permutations to compute a pvalue for DSC using a permutation test}

\item{digits.d}{number of decimal places for DSC}

\item{digits.p}{number of decimal places for p-value}
}
\value{
A list with elements
\item{DSC}{Dispersion Separability Criterion}
\item{pval}{Permutation test p-value}
}
\description{
The DSC indicates how similar two data sets are by considering
the between data scatter matrix and the within data scatter matrix.
}
\details{
First, the different batches in \code{data} are combined into a single dataset. Under
the null hypothesis, there is no difference between the two groups. We sample from the
combined dataset, then compute the null (permuted) difference.
The p-value in the permutation test is the fraction of times the permuted
difference is equal to or more extreme than the observed difference.
}
\examples{
set.seed(4)
dat <- matrix(rnorm(100), ncol = 10)
batches <- sample(1:2, 10, replace = TRUE)
DSC(dat, batches, permut = 1000)
}
\author{
Aline Talhouk
}

