% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mycombat_p.R
\name{mycombat_p}
\alias{mycombat_p}
\title{Custom combat_p}
\usage{
mycombat_p(Y, b, prior.plots = F)
}
\arguments{
\item{Y}{To be filled.}

\item{b}{To be filled.}

\item{prior.plots}{To be filled.}
}
\description{
Based on combat_p function from package \code{ber}
}

