% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scattermat.R
\name{scattermat}
\alias{scattermat}
\title{Scatter Matrices}
\usage{
scattermat(data, Y)
}
\arguments{
\item{data}{data matrix}

\item{Y}{vector of classes}
}
\value{
A list with the following elements
\item{B}{between class scatter matrix}
\item{W}{within class scatter matrix}
}
\description{
Computes the Between and Within class scatter matrices.
}
\details{
The vector \code{Y} can indicate distinct classes, groups, batches, etc. Between and
within class scatter matrices are computed based on the indices of \code{Y}.
}
\examples{
set.seed(4)
dat <- matrix(rnorm(100), ncol = 10)
batches <- sample(1:2, 10, replace = TRUE)
scattermat(dat, batches)
}
\author{
Aline Talhouk
}

