% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scattermat.R
\name{scattermat}
\alias{scattermat}
\title{Computes the between and within class scatter matrices}
\usage{
scattermat(data, Y)
}
\arguments{
\item{data}{data matrix}

\item{Y}{vector of classes}
}
\value{
A list with the following elements
\item{B}{between class scatter matrix}
\item{W}{within class scatter matrix}
}
\description{
Computes the between and within class scatter matrices
}
\examples{
set.seed(4)
dat <- matrix(rnorm(100), ncol = 10)
batches <- sample(1:2, 10, replace = TRUE)
scattermat(dat, batches)
}
\author{
Aline Talhouk
}

